Added more functionality to ExceptionHandler and FileHelper according to the comments.

+++ src/Utils/FileHelper.cs
@@ -4,10 +4,12 @@ static class FileHelper
 {
     public static string[] ReadFile(string filePath)
     {
+        if (!File.Exists(filePath)) throw ExceptionHandler.FileNotFound();
         return File.ReadAllLines(filePath);
     }
     public static void WriteFile(string filePath, string[] data)
     {
+        if (!File.Exists(filePath)) throw ExceptionHandler.FileNotFound();
         File.WriteAllLines(filePath, data);
     }
 }
